+++ studio/constants/constants.js
@@ -1,0 +1,3 @@
+export const constants = {
+  siteUrl: 'https://imsety.com'
+};